pyproject.toml
[project]
name = "img-to-svg"
version = "0.1.0"
description = "Convert JPG/PNG images into vector SVG with color or black-and-white modes"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=2.0.0",
  "pillow>=10.0.0",
  "scikit-image>=0.24.0",
  "svgwrite>=1.4.3",
  "vtracer>=0.6.11",
]

[project.scripts]
img-to-svg = "img_to_vector:main"

[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["img_to_vector.py"]

[tool.hatch.build.targets.sdist]
include = ["img_to_vector.py", "README.md", "pyproject.toml"]
